+++ ts/tests/ser.test.ts
@@ -3,26 +3,26 @@ import * as schm from "@massage/Test";
 
 test("serialize empty", () => {
   const m = new schm.Test(new schm.Empty());
-  expect(() => m.serialize()).not.toThrow();
+  expect(() => m._serialize()).not.toThrow();
 });
 
 test("serialize u8", () => {
   const m = new schm.Test(new schm.U8(69));
-  const ser = m.serialize();
+  const ser = m._serialize();
   expect(ser[1]).toBe(69);
   expect(ser.length).toBe(2);
 });
 
 test("serialize bool", () => {
   const m = new schm.Test(new schm.Bool(true));
-  const ser = m.serialize();
+  const ser = m._serialize();
   expect(ser[1]).toBe(1);
   expect(ser.length).toBe(2);
 });
 
 test("serialize u16", () => {
   const m = new schm.Test(new schm.U16(0x1234));
-  const ser = m.serialize();
+  const ser = m._serialize();
   expect(ser[1]).toBe(0x34);
   expect(ser[2]).toBe(0x12);
   expect(ser.length).toBe(3);
@@ -30,7 +30,7 @@ test("serialize u16", () => {
 
 test("serialize u32", () => {
   const m = new schm.Test(new schm.U32(0x12345678));
-  const ser = m.serialize();
+  const ser = m._serialize();
   expect(ser[1]).toBe(0x78);
   expect(ser[2]).toBe(0x56);
   expect(ser[3]).toBe(0x34);
@@ -40,7 +40,7 @@ test("serialize u32", () => {
 
 test("serialize u64", () => {
   const m = new schm.Test(new schm.U64(0x123456789abcdef0n));
-  const ser = m.serialize();
+  const ser = m._serialize();
   expect(ser[1]).toBe(0xf0);
   expect(ser[2]).toBe(0xde);
   expect(ser[3]).toBe(0xbc);
@@ -54,14 +54,14 @@ test("serialize u64", () => {
 
 test("serialize i8", () => {
   const m = new schm.Test(new schm.I8(69));
-  const ser = m.serialize();
+  const ser = m._serialize();
   expect(ser[1]).toBe(69);
   expect(ser.length).toBe(2);
 });
 
 test("serialize i16", () => {
   const m = new schm.Test(new schm.I16(0x1234));
-  const ser = m.serialize();
+  const ser = m._serialize();
   expect(ser[1]).toBe(0x34);
   expect(ser[2]).toBe(0x12);
   expect(ser.length).toBe(3);
@@ -69,7 +69,7 @@ test("serialize i16", () => {
 
 test("serialize i32", () => {
   const m = new schm.Test(new schm.I32(0x12345678));
-  const ser = m.serialize();
+  const ser = m._serialize();
   expect(ser[1]).toBe(0x78);
   expect(ser[2]).toBe(0x56);
   expect(ser[3]).toBe(0x34);
@@ -79,7 +79,7 @@ test("serialize i32", () => {
 
 test("serialize i64", () => {
   const m = new schm.Test(new schm.I64(0x123456789abcdef0n));
-  const ser = m.serialize();
+  const ser = m._serialize();
   expect(ser[1]).toBe(0xf0);
   expect(ser[2]).toBe(0xde);
   expect(ser[3]).toBe(0xbc);
@@ -93,14 +93,14 @@ test("serialize i64", () => {
 
 test("serialize i8 negative", () => {
   const m = new schm.Test(new schm.I8(-69));
-  const ser = m.serialize();
+  const ser = m._serialize();
   expect(ser[1]).toBe(187);
   expect(ser.length).toBe(2);
 });
 
 test("serialize i16 negative", () => {
   const m = new schm.Test(new schm.I16(-0x1234));
-  const ser = m.serialize();
+  const ser = m._serialize();
   expect(ser[1]).toBe(0xcc);
   expect(ser[2]).toBe(0xed);
   expect(ser.length).toBe(3);
@@ -108,7 +108,7 @@ test("serialize i16 negative", () => {
 
 test("serialize i32 negative", () => {
   const m = new schm.Test(new schm.I32(-0x12345678));
-  const ser = m.serialize();
+  const ser = m._serialize();
   expect(ser[1]).toBe(0x88);
   expect(ser[2]).toBe(0xa9);
   expect(ser[3]).toBe(0xcb);
@@ -118,7 +118,7 @@ test("serialize i32 negative", () => {
 
 test("serialize i64 negative", () => {
   const m = new schm.Test(new schm.I64(-0x123456789abcdef0n));
-  const ser = m.serialize();
+  const ser = m._serialize();
   expect(ser[1]).toBe(0x10);
   expect(ser[2]).toBe(0x21);
   expect(ser[3]).toBe(0x43);
@@ -132,7 +132,7 @@ test("serialize i64 negative", () => {
 
 test("serialize string", () => {
   const m = new schm.Test(new schm.Str("hello!"));
-  const ser = m.serialize();
+  const ser = m._serialize();
   expect(ser[1]).toBe(0x68);
   expect(ser[7]).toBe(0x0);
   expect(ser.length).toBe(8);
@@ -140,7 +140,7 @@ test("serialize string", () => {
 
 test("serialize f32", () => {
   const m = new schm.Test(new schm.F32(Math.PI));
-  const ser = m.serialize();
+  const ser = m._serialize();
   expect(ser[1]).toBe(0xdb);
   expect(ser[2]).toBe(0x0f);
   expect(ser[3]).toBe(0x49);
@@ -150,7 +150,7 @@ test("serialize f32", () => {
 
 test("serialize f64", () => {
   const m = new schm.Test(new schm.F64(Math.PI));
-  const ser = m.serialize();
+  const ser = m._serialize();
   expect(ser[1]).toBe(0x18);
   expect(ser[2]).toBe(0x2d);
   expect(ser[3]).toBe(0x44);
@@ -164,7 +164,7 @@ test("serialize f64", () => {
 
 test("serialize zero termination", () => {
   const m = new schm.Test(new schm.ZeroTermination("hello!", 69));
-  const ser = m.serialize();
+  const ser = m._serialize();
   expect(ser[7]).toBe(0x0);
   expect(ser.length).toBe(9);
 });